+++ packages/components-react/projects/nextjs/routes.ts
@@ -236,9 +236,13 @@ const generatedRoutes = {
     "path": "/flyout-basic",
     "name": "FlyoutBasic"
   },
-  "flyoutCssVariables": {
-    "path": "/flyout-css-variables",
-    "name": "FlyoutCssVariables"
+  "flyoutCssVariables_1": {
+    "path": "/flyout-css-variables-1",
+    "name": "FlyoutCssVariables_1"
+  },
+  "flyoutCssVariables_2": {
+    "path": "/flyout-css-variables-2",
+    "name": "FlyoutCssVariables_2"
   },
   "flyoutFixedFooter": {
     "path": "/flyout-fixed-footer",
@@ -563,6 +567,10 @@ const generatedRoutes = {
   "popoverBasic": {
     "path": "/popover-basic",
     "name": "PopoverBasic"
+  },
+  "popoverBasicSlotted": {
+    "path": "/popover-basic-slotted",
+    "name": "PopoverBasicSlotted"
   },
   "popoverBottom": {
     "path": "/popover-bottom",